refactor: 사용자 프로필 조회 및 팔로우 관련 src/shared로 이관
